Roles

+++ app/Http/Controllers/CourseController.php
@@ -2,14 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\FilterByAreaRequest;
 use App\Http\Requests\FilterRequest;
 use App\Http\Requests\StoreCourseRequest;
 use App\Http\Requests\UpdateCourseRequest;
+use App\Models\Area;
 use App\Models\Course;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\Response;
-use Illuminate\Http\Request;
+use Silber\Bouncer\BouncerFacade as Bouncer;
 
 class CourseController extends Controller
 {
@@ -18,7 +17,27 @@ class CourseController extends Controller
      */
     public function index(): Response
     {
-        return response(Course::all(), 200);
+        $orderedByRating = Course::query()->orderBy('rating', 'desc')->get();
+
+        return response($orderedByRating->groupBy('area_id'), 200);
+    }
+
+    /**
+     * Apply all the filters
+     */
+    public function filter(FilterRequest $request): Response
+    {
+        $request->validated();
+
+        $this->checkIfAreaExist($request['area_id']);
+
+        $filtered = Course::query()
+            ->where('area_id', '=', $request['area_id'])
+            ->where('price', '>=', $request['minPrice'])
+            ->where('price', '<=', $request['maxPrice'])
+            ->orderBy('price')->get();
+
+        return \response($filtered, 200);
     }
 
     /**
@@ -27,16 +46,26 @@ class CourseController extends Controller
     public function store(StoreCourseRequest $request): Response
     {
         $request->validated();
-        $course = Course::query()->firstOrCreate($request->all());
+
+        $user = $this->checkInstructorRole();
+
+        $this->checkIfAreaExist($request['area_id']);
+
+        $course = $user->courses()->firstOrCreate($request->all());
+
+        $course->available_places = $course->max_students;
+
+        $course->save();
+
         return response($course, 201);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Course $course): Course
+    public function show(Course $course): Response
     {
-        return $course;
+        return \response($course->load(['ratings', 'user']), 200);
     }
 
     /**
@@ -44,7 +73,19 @@ class CourseController extends Controller
      */
     public function update(UpdateCourseRequest $request, Course $course): Response
     {
-        return $course->update($request->all()) ? \response($course) : \response(null, 400);
+        $request->validated();
+
+        $user = $this->checkInstructorRole();
+
+        if ($user->id !== $course->user_id) {
+            abort(403, 'Solo los propietarios del curso pueden editarlo');
+        }
+
+        $this->checkIfAreaExist($request['area_id']);
+
+        $course->update($request->all());
+
+        return \response($course, 200);
     }
 
     /**
@@ -52,30 +93,30 @@ class CourseController extends Controller
      */
     public function destroy(Course $course): Response
     {
+        $user = $this->checkInstructorRole();
+
+        if ($user->id !== $course->user_id) {
+            abort(403, 'Solo los propietarios del curso pueden eliminarlo');
+        }
+
         $course->delete();
+
         return \response(null, 204);
     }
 
-    /**
-     * Apply all the filters
-     */
-    public function filter(FilterRequest $request): Response
+    public function checkInstructorRole()
     {
-        $criterias = $request->validated();
+        $user = request()->user();
 
-        $filtered = Course::query()
-            ->where('area', '=', $criterias['area'])
-            ->where('price', '<=', $criterias['price'])->get();
+        if (Bouncer::is($user)->notAn('instructor')) {
+            abort(403, 'Solo instructores pueden gestionar los cursos');
+        }
 
-        return \response($filtered, 200);
+        return $user;
     }
 
-    /**
-     * Filter courses by area
-     */
-    public function filterByArea($areaCriteria): Collection
+    public function checkIfAreaExist($areaId)
     {
-        return Course::query()->where('area', '=', $areaCriteria)->get();
+        Area::query()->findOrFail($areaId);
     }
-
 }
